backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.9.0
pydantic-settings==2.5.0
python-dotenv==1.0.1

# Azure SDKs
azure-search-documents==11.6.0
azure-cosmos==4.7.0
azure-storage-blob==12.22.0
azure-identity==1.17.0
google-genai==1.0.0

# File generation
python-pptx==1.0.2
openpyxl==3.1.5

# Utilities
httpx==0.27.0
sse-starlette==2.1.0
